Face-Emotion-Recognition-main/requirements.txt
# core
numpy>=1.23.5,<2
pandas>=1.5.3
scikit-learn>=1.2.2
scipy>=1.9.3
matplotlib>=3.6.0
pillow>=10.0.0

# TF/Keras (compatible pins)
tensorflow-cpu==2.12.1
keras==2.12.0
keras-preprocessing==1.1.2
opt-einsum==3.3.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
typing-extensions>=4.5.0,<4.6.0

# OpenCV (GUI windows)
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Flask API (if you want the local endpoint)
flask==3.0.3
flask-cors==4.0.0
werkzeug==3.0.3
jinja2==3.1.3
blinker==1.7.0
itsdangerous>=2.1.2

# misc utilities
protobuf==4.24.0
PyYAML==6.0.2
cffi==1.15.1
argon2-cffi==21.3.0
cryptography==41.0.0
tornado==6.3.3
requests==2.31.0
urllib3==2.0.7
cachetools==5.3.2
rsa==4.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
attrs==23.1.0
packaging==23.2
toml==0.10.2
toolz==0.12.0
six==1.16.0
